[v2-api] separate JSR from Engine
